% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{factorial}
\alias{factorial}
\title{Factorial means}
\usage{
data(factorial)
}
\description{
Simulated data testing factorial power analysis based on means
}
\examples{
data(factorial)
}
\keyword{datasets}
