% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pamlcorr.f.R
\name{pamlcorr}
\alias{pamlcorr}
\title{Correlation}
\usage{
pamlcorr(
  aim = "n",
  es = 0.5,
  power = 0.9,
  n = 20,
  sig.level = 0.05,
  alternative = "two.sided",
  plot_contour = FALSE,
  plot_escurve = FALSE,
  plot_ncurve = FALSE,
  plot_log = TRUE,
  plot_palette = "viridis",
  plot_x = "none",
  plot_y = "none",
  plot_custom_labels = FALSE,
  plot_z = "none",
  plot_x_from = 0,
  plot_x_to = 0,
  plot_z_lines = 1,
  plot_z_value = list(),
  plot_to_table = FALSE,
  .interface = "R",
  .caller = "correlation"
)
}
\arguments{
\item{aim}{The aim of the analysis: \code{n} (default) for sample size,
\code{power} to estimate power, \code{es} for effect size (correlation)}

\item{es}{The expected effect size, i.e. the expected correlation}

\item{power}{Minimal desired power}

\item{n}{Sample size}

\item{sig.level}{Type I error rate (significance cut-off or alpha)}

\item{alternative}{`one.side` or `two.sided` (default)}

\item{plot_contour}{Produce power contour plot}

\item{plot_escurve}{Produce power as a function of possible effect size values plot}

\item{plot_ncurve}{Produce power as a function of possible sample size values plot}

\item{plot_log}{.   Use log-scale in plots}

\item{plot_palette}{Change the palettes of the plots}

\item{plot_x}{Parameter in the X-axis of custom power parameters plot.
\code{n}, \code{power} or \code{es}, the latest being the effect size.}

\item{plot_y}{Parameter in the Y-axis of custom power parameters plot.
\code{n}, \code{power} or \code{es}, the latest being the effect size.}

\item{plot_custom_labels}{Add value labels to the custom power parameters
plot, default FALSE.}

\item{plot_z}{break down the custom power parameters plot by a parameter.
\code{n}, \code{power}, \code{sig.level} or \code{es} for effect size}

\item{plot_x_from}{Range for the parameter in the X-axis of custom power
parameters plot. Starting value.}

\item{plot_x_to}{Range for the parameter in the X-axis of custom power
parameters plot. Ending value.}

\item{plot_z_lines}{not used in R}

\item{plot_z_value}{Values of the parameter to break the plot by}

\item{plot_to_table}{Produce a table of the values plotted in custom power
parameters plot}

\item{.interface}{Used for internal purposes}

\item{.caller}{Used for internal purposes}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$intro} \tab \tab \tab \tab \tab a html \cr
  \code{results$extrainfo} \tab \tab \tab \tab \tab a html \cr
  \code{results$issues} \tab \tab \tab \tab \tab a html \cr
  \code{results$powertab} \tab \tab \tab \tab \tab a table \cr
  \code{results$powerbyes} \tab \tab \tab \tab \tab a table \cr
  \code{results$plotnotes} \tab \tab \tab \tab \tab a html \cr
  \code{results$powerContour} \tab \tab \tab \tab \tab an image \cr
  \code{results$powerEscurve} \tab \tab \tab \tab \tab an image \cr
  \code{results$powerNcurve} \tab \tab \tab \tab \tab an image \cr
  \code{results$powerCustom} \tab \tab \tab \tab \tab an image \cr
  \code{results$customnotes} \tab \tab \tab \tab \tab a html \cr
  \code{results$customtable} \tab \tab \tab \tab \tab a table \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$powertab$asDF}

\code{as.data.frame(results$powertab)}
}
\description{
Compute power parameters for Pearson correlations
}
