% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pamlfactorial.h.R
\name{pamlfactorial}
\alias{pamlfactorial}
\title{Factorial Designs}
\usage{
pamlfactorial(
  data,
  .caller = "factorial",
  .interface = "jamovi",
  aim = "n",
  mode = "facpeta",
  means = NULL,
  sds = NULL,
  factors = NULL,
  within = NULL,
  r = 0,
  peta = 0.2,
  effect_type,
  repeated_type,
  df_effect = 1,
  effect_groups = 2,
  design_groups = 2,
  power = 0.9,
  n = 20,
  sig.level = 0.05,
  plot_contour = FALSE,
  plot_escurve = FALSE,
  plot_ncurve = FALSE,
  plot_log = FALSE,
  plot_x = "none",
  plot_y = "none",
  plot_custom_labels = FALSE,
  plot_z = "none",
  plot_x_from = 0,
  plot_x_to = 0,
  plot_z_lines = 0,
  plot_z_value = list(),
  plot_to_table = FALSE,
  ncp_type,
  emeans = FALSE,
  esos = FALSE
)
}
\arguments{
\item{data}{the data as a data frame}

\item{.caller}{.}

\item{.interface}{.}

\item{aim}{The aim of the analysis: \code{n} (default) for sample size,
\code{power} to estimate power}

\item{mode}{.}

\item{means}{a string naming the variable containing the means from
\code{data}; the variable must be numeric.}

\item{sds}{a string naming the variable containing the standard deviations
from \code{data}; the variable must be numeric.}

\item{factors}{a vector of strings naming the  factors from \code{data}.
Not needed if \code{formula} is used.}

\item{within}{a vector of strings naming the repeated measures factors from
\code{data}.}

\item{r}{Expected correlation among repeated measures}

\item{peta}{.}

\item{effect_type}{.}

\item{repeated_type}{.}

\item{df_effect}{.}

\item{effect_groups}{.}

\item{design_groups}{.}

\item{power}{Minimal desired power}

\item{n}{Sample size}

\item{sig.level}{Type I error rate (significance cut-off or alpha)}

\item{plot_contour}{.}

\item{plot_escurve}{.}

\item{plot_ncurve}{.}

\item{plot_log}{.}

\item{plot_x}{.}

\item{plot_y}{.}

\item{plot_custom_labels}{.}

\item{plot_z}{.}

\item{plot_x_from}{.}

\item{plot_x_to}{.}

\item{plot_z_lines}{.}

\item{plot_z_value}{.}

\item{plot_to_table}{.}

\item{ncp_type}{.}

\item{emeans}{.}

\item{esos}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$intro} \tab \tab \tab \tab \tab a html \cr
  \code{results$issues} \tab \tab \tab \tab \tab a html \cr
  \code{results$powertab} \tab \tab \tab \tab \tab a table \cr
  \code{results$powerbyes} \tab \tab \tab \tab \tab a table \cr
  \code{results$means} \tab \tab \tab \tab \tab an array of predicted means tables \cr
  \code{results$plotnotes} \tab \tab \tab \tab \tab a html \cr
  \code{results$powerContour} \tab \tab \tab \tab \tab an image \cr
  \code{results$powerEscurve} \tab \tab \tab \tab \tab an image \cr
  \code{results$powerNcurve} \tab \tab \tab \tab \tab an image \cr
  \code{results$powerCustom} \tab \tab \tab \tab \tab an image \cr
  \code{results$customnotes} \tab \tab \tab \tab \tab a html \cr
  \code{results$customtable} \tab \tab \tab \tab \tab a table \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$powertab$asDF}

\code{as.data.frame(results$powertab)}
}
\description{
Something here
}
