% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pamlprop.h.R
\name{pamlprop}
\alias{pamlprop}
\title{Proportions}
\usage{
pamlprop(
  .caller = "proportions",
  aim = "n",
  mode = "propind",
  propind_p1 = 0.6,
  propind_p2 = 0.5,
  propind_n = 20,
  propind_nratio = 1,
  propone_p1 = 0.6,
  propone_p2 = 0.5,
  propone_n = 50,
  proppaired_p1 = 0.1,
  proppaired_p2 = 0.4,
  proppaired_n = 20,
  power = 0.9,
  sig.level = 0.05,
  alternative = "two.sided",
  plot_contour = FALSE,
  plot_escurve = FALSE,
  plot_ncurve = FALSE,
  plot_log = TRUE,
  plot_x = "none",
  plot_y = "none",
  plot_custom_labels = FALSE,
  plot_z = "none",
  plot_x_from = 0,
  plot_x_to = 0,
  plot_z_lines = 0,
  plot_z_value = list(),
  plot_to_table = FALSE,
  es_type = "odd"
)
}
\arguments{
\item{.caller}{.}

\item{aim}{.}

\item{mode}{.}

\item{propind_p1}{.}

\item{propind_p2}{.}

\item{propind_n}{.}

\item{propind_nratio}{.}

\item{propone_p1}{.}

\item{propone_p2}{.}

\item{propone_n}{.}

\item{proppaired_p1}{.}

\item{proppaired_p2}{.}

\item{proppaired_n}{.}

\item{power}{.}

\item{sig.level}{.}

\item{alternative}{.}

\item{plot_contour}{.}

\item{plot_escurve}{.}

\item{plot_ncurve}{.}

\item{plot_log}{.}

\item{plot_x}{.}

\item{plot_y}{.}

\item{plot_custom_labels}{.}

\item{plot_z}{.}

\item{plot_x_from}{.}

\item{plot_x_to}{.}

\item{plot_z_lines}{.}

\item{plot_z_value}{.}

\item{plot_to_table}{.}

\item{es_type}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$intro} \tab \tab \tab \tab \tab a html \cr
  \code{results$issues} \tab \tab \tab \tab \tab a html \cr
  \code{results$powertab} \tab \tab \tab \tab \tab a table \cr
  \code{results$powerbyes} \tab \tab \tab \tab \tab a table \cr
  \code{results$plotnotes} \tab \tab \tab \tab \tab a html \cr
  \code{results$powerContour} \tab \tab \tab \tab \tab an image \cr
  \code{results$powerEscurve} \tab \tab \tab \tab \tab an image \cr
  \code{results$powerNcurve} \tab \tab \tab \tab \tab an image \cr
  \code{results$powerCustom} \tab \tab \tab \tab \tab an image \cr
  \code{results$customnotes} \tab \tab \tab \tab \tab a html \cr
  \code{results$customtable} \tab \tab \tab \tab \tab a table \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$powertab$asDF}

\code{as.data.frame(results$powertab)}
}
\description{
Something here
}
