% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pamlsem.h.R
\name{pamlsem}
\alias{pamlsem}
\title{Structural Equations Models}
\usage{
pamlsem(
  aim = "n",
  code = "",
  fonts = "small",
  toggle = FALSE,
  power = 0.9,
  n = 100,
  sig.level = 0.05,
  alternative = "two.sided",
  estimator = "ML",
  standardized = TRUE,
  inspect_sigma = FALSE,
  inspect_lvcov = FALSE,
  inspect_zbeta = FALSE,
  method = "analytic",
  mc_test = "lrt",
  mcR = 500,
  parallel = TRUE,
  set_seed = FALSE,
  seed = 42,
  table_pwbyn = TRUE,
  plot_ncurve = FALSE,
  plot_log = FALSE,
  plot_palette = "viridis",
  plot_x = "none",
  plot_y = "none",
  plot_custom_labels = FALSE,
  plot_z = "none",
  plot_x_from = 0,
  plot_x_to = 0,
  plot_z_lines = 1,
  plot_z_value = list(),
  plot_to_table = FALSE,
  explain = FALSE,
  lav_diagram = TRUE,
  diag_size = "medium",
  .interface = "jamovi",
  .caller = "pamlsem"
)
}
\arguments{
\item{aim}{The aim of the analysis: \code{n} (default) for sample size,
\code{power} to estimate power}

\item{code}{.}

\item{fonts}{.}

\item{toggle}{.}

\item{power}{Minimal desired power}

\item{n}{Sample size}

\item{sig.level}{Type I error rate (significance cut-off or alpha)}

\item{alternative}{Two-tailed vs one-tailed test}

\item{estimator}{Estimator method used in SEM. \code{ML} for Maximum
Likelihood}

\item{standardized}{Assume the observed variables are standardized (TRUE)
or not (FALSE)}

\item{inspect_sigma}{not used in R}

\item{inspect_lvcov}{not used in R}

\item{inspect_zbeta}{not used in R}

\item{method}{Use analytic methods for computation of power parameter
(\code{analytic}) or Monte Carlo (\code{mc})}

\item{mc_test}{which test is used in Monte Carlo simulations: \code{lrt}
for LRT or \code{score} for the Score test.}

\item{mcR}{Number of repetitions for Monte Carlo method}

\item{parallel}{Logical: should parallel computing be used for the Monte
Carlo method}

\item{set_seed}{not used in R}

\item{seed}{not used in R}

\item{table_pwbyn}{.}

\item{plot_ncurve}{.}

\item{plot_log}{.}

\item{plot_palette}{.}

\item{plot_x}{.}

\item{plot_y}{.}

\item{plot_custom_labels}{.}

\item{plot_z}{.}

\item{plot_x_from}{.}

\item{plot_x_to}{.}

\item{plot_z_lines}{.}

\item{plot_z_value}{.}

\item{plot_to_table}{.}

\item{explain}{not used in R}

\item{lav_diagram}{Output the path diagram corresponding to the input model}

\item{diag_size}{Size of the path diagram boxes: \code{small},
\code{medium}, \code{large}}

\item{.interface}{Used for internal purposes}

\item{.caller}{Used for internal purposes}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$intro} \tab \tab \tab \tab \tab a html \cr
  \code{results$extrainfo} \tab \tab \tab \tab \tab a html \cr
  \code{results$issues} \tab \tab \tab \tab \tab a html \cr
  \code{results$initnotes} \tab \tab \tab \tab \tab a html \cr
  \code{results$diagram} \tab \tab \tab \tab \tab an image \cr
  \code{results$powertab} \tab \tab \tab \tab \tab a table \cr
  \code{results$powerbyn} \tab \tab \tab \tab \tab a table \cr
  \code{results$implied$covs} \tab \tab \tab \tab \tab a table \cr
  \code{results$implied$lvcovs} \tab \tab \tab \tab \tab a table \cr
  \code{results$implied$betas} \tab \tab \tab \tab \tab a table \cr
  \code{results$plotnotes} \tab \tab \tab \tab \tab a html \cr
  \code{results$powerNcurve} \tab \tab \tab \tab \tab an image \cr
  \code{results$powerCustom} \tab \tab \tab \tab \tab an image \cr
  \code{results$customnotes} \tab \tab \tab \tab \tab a html \cr
  \code{results$customtable} \tab \tab \tab \tab \tab a table \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$powertab$asDF}

\code{as.data.frame(results$powertab)}
}
\description{
Something here
}
