% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pamlttest.h.R
\name{pamlttest}
\alias{pamlttest}
\title{T-Test}
\usage{
pamlttest(
  aim = "n",
  mode = "ttestind",
  ttestind_es = 0.5,
  ttestind_n = 20,
  ttestind_nratio = 1,
  ttestpaired_es = 0.5,
  ttestpaired_n = 20,
  ttestone_es = 0.5,
  ttestone_n = 20,
  is_equi = FALSE,
  equi_limit = 0.1,
  power = 0.9,
  sig.level = 0.05,
  alternative = "two.sided",
  plot_contour = FALSE,
  plot_escurve = FALSE,
  plot_ncurve = FALSE,
  plot_log = TRUE,
  plot_palette = "viridis",
  plot_x = "none",
  plot_y = "none",
  plot_custom_labels = FALSE,
  plot_z = "none",
  plot_x_from = 0,
  plot_x_to = 0,
  plot_z_lines = 0,
  plot_z_value = list(),
  plot_to_table = FALSE,
  explain = FALSE,
  .caller = "ttest",
  .interface = "jamovi",
  ttestind_ttest = "-",
  ttestind_ngroup = 0,
  ttestind_d = 0,
  ttestind_useit,
  ttestpaired_ttest = "-",
  ttestpaired_obsn = 0,
  ttestpaired_d = 0,
  ttestpaired_useit,
  ttestpaired_dind = "-",
  ttestpaired_r = 0,
  ttestpaired_dz = 0,
  ttestpaired_d_useit,
  ttestone_ttest = "-",
  ttestone_obsn = 0,
  ttestone_d = 0,
  ttestone_useit
)
}
\arguments{
\item{aim}{The aim of the analysis: \code{n} for required N, \code{power}
for estimated power and \code{es} for minimal detectable effect size}

\item{mode}{.}

\item{ttestind_es}{Expected Cohe's d}

\item{ttestind_n}{Sample size per group}

\item{ttestind_nratio}{Relative size of the two groups. 1 means group of
equal size, \code{ratio}>1 means that one group is \code{ratio} times
larger than the other.}

\item{ttestpaired_es}{Expected Cohe's d_z, defined as the mean difference
divided by the standard deviation of the  difference score}

\item{ttestpaired_n}{Expected sample size (number of paires)}

\item{ttestone_es}{Expected Cohe's d}

\item{ttestone_n}{Expected sample size}

\item{is_equi}{When \code{TRUE}, perform an equivalence test to show that a
true effect is small enough to be considered practically negligible}

\item{equi_limit}{.}

\item{power}{Minimal desired power (1-Beta)}

\item{sig.level}{Critical alpha (significance level to be used in the
analysis)}

\item{alternative}{Wheather the test is two-tailed (\code{two.sided}) or
one-tailed (\code{greater}).}

\item{plot_contour}{.}

\item{plot_escurve}{.}

\item{plot_ncurve}{.}

\item{plot_log}{.}

\item{plot_palette}{.}

\item{plot_x}{.}

\item{plot_y}{.}

\item{plot_custom_labels}{.}

\item{plot_z}{.}

\item{plot_x_from}{.}

\item{plot_x_to}{.}

\item{plot_z_lines}{.}

\item{plot_z_value}{.}

\item{plot_to_table}{.}

\item{explain}{not used in R}

\item{.caller}{.}

\item{.interface}{.}

\item{ttestind_ttest}{not used in R}

\item{ttestind_ngroup}{not used in R}

\item{ttestind_d}{not used in R}

\item{ttestind_useit}{not used in R}

\item{ttestpaired_ttest}{not used in R}

\item{ttestpaired_obsn}{not used in R}

\item{ttestpaired_d}{not used in R}

\item{ttestpaired_useit}{not used in R}

\item{ttestpaired_dind}{not used in R}

\item{ttestpaired_r}{not used in R}

\item{ttestpaired_dz}{not used in R}

\item{ttestpaired_d_useit}{not used in R}

\item{ttestone_ttest}{not used in R}

\item{ttestone_obsn}{not used in R}

\item{ttestone_d}{not used in R}

\item{ttestone_useit}{not used in R}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$intro} \tab \tab \tab \tab \tab a html \cr
  \code{results$issues} \tab \tab \tab \tab \tab a html \cr
  \code{results$extrainfo} \tab \tab \tab \tab \tab a html \cr
  \code{results$powertab} \tab \tab \tab \tab \tab a table \cr
  \code{results$powerbyes} \tab \tab \tab \tab \tab a table \cr
  \code{results$plotnotes} \tab \tab \tab \tab \tab a html \cr
  \code{results$powerContour} \tab \tab \tab \tab \tab an image \cr
  \code{results$powerEscurve} \tab \tab \tab \tab \tab an image \cr
  \code{results$powerNcurve} \tab \tab \tab \tab \tab an image \cr
  \code{results$powerCustom} \tab \tab \tab \tab \tab an image \cr
  \code{results$customnotes} \tab \tab \tab \tab \tab a html \cr
  \code{results$customtable} \tab \tab \tab \tab \tab a table \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$powertab$asDF}

\code{as.data.frame(results$powertab)}
}
\description{
Something here
}
