% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rinterface.R
\name{plots}
\alias{plots}
\alias{plot.pamlj}
\title{pamlj plots}
\usage{
plots(x, ...)

\method{plot}{pamlj}(x, ...)
}
\arguments{
\item{x}{a pamlj results object of the class `pamlj`}
}
\value{
an object of class ggplot or a list of ggplot objects

an object of class ggplot or a list of ggplot objects
}
\description{
This function extract plots from a pamlj object and return them as a list

This function extract plots from a pamlj object.
}
\examples{
obj<-pamlj::pamlcorr(es=.1,.interface="R",plot_contour=T,plot_x="n",plot_y="es",plot_x_from=10,plot_x_to=100)
results<-plots(obj)
lapply(results,class)
\dontrun{
obj<-pamlj::pamlcorr(es=.1,.interface="R",plot_contour=T,plot_x="n",plot_y="es",plot_x_from=10,plot_x_to=100)
plot(obj)
}
}
\author{
Marcello Gallucci
}
