% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rinterface.R
\name{summary.ResultsElement}
\alias{summary.ResultsElement}
\alias{summary.pamlj_list}
\alias{print.jmvrtable}
\alias{print.jmvrobj}
\alias{print.pamlj_list}
\title{S3 methods for class jamovi ResultsElement}
\usage{
\method{summary}{ResultsElement}(object, ...)

\method{summary}{pamlj_list}(object, ...)

\method{print}{jmvrtable}(x, ...)

\method{print}{jmvrobj}(x, ...)

\method{print}{pamlj_list}(x, ...)
}
\arguments{
\item{object}{a pamlj results object of the class `pamlj`}

\item{...}{options passed to print()}

\item{x}{a pamlj results object of the class `pamlj`}
}
\value{
a list of table as data.frame

a list of tables as data.frame
}
\description{
These functions extract all visible tables from a ResultsElement or related classes produced by pamlj
and print them in R style.

These functions extract all visible tables from a list of tables produced by pamlj3
and print them in R style.
}
\examples{
obj<-pamlj::pamlcorr(es=.1)
summary(obj)
data(fivegroups)
fivegroups$Group<-factor(fivegroups$Group)
gmod<-pamlj3::pamlj_lm(
  formula = Score ~Group,
  data = fivegroups)

summary(gmod)
}
\author{
Marcello Gallucci
}
