% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pamlglm.h.R
\name{pamlglm}
\alias{pamlglm}
\title{General Linear Model}
\usage{
pamlglm(
  data,
  aim = "n",
  mode = "peta",
  b_es = 0.2,
  b_r2 = 0.05,
  b_df_model = 1,
  v_es = 0.2,
  v_df_model = 1,
  v_df_effect = 1,
  e_es = 0.2,
  e_df_model = 1,
  e_r2 = 0.2,
  e_df_effect = 1,
  power = 0.9,
  n = 20,
  sig.level = 0.05,
  alternative = "two.sided",
  plot_contour = FALSE,
  plot_escurve = FALSE,
  plot_ncurve = FALSE,
  plot_log = FALSE,
  plot_palette = "viridis",
  covs = 0,
  factors = 0,
  factors_list = list(list(var = "factor 1", levels = 0)),
  covs_order,
  factors_order,
  mixed_order,
  eta = "-",
  f = "-",
  eta_df_error = 0,
  epsilon = 0,
  omega = 0,
  gpower = 0,
  f2 = 0,
  use = "none",
  ncp_type,
  rx = NULL,
  plot_x = "none",
  plot_y = "none",
  plot_custom_labels = FALSE,
  plot_z = "none",
  plot_x_from = 0,
  plot_x_to = 0,
  plot_z_lines = 0,
  plot_z_value = list(),
  plot_to_table = FALSE,
  explain = FALSE,
  .caller = "glm",
  .interface = "jamovi"
)
}
\arguments{
\item{data}{the data as a data frame}

\item{aim}{.}

\item{mode}{.}

\item{b_es}{.}

\item{b_r2}{.}

\item{b_df_model}{.}

\item{v_es}{.}

\item{v_df_model}{.}

\item{v_df_effect}{.}

\item{e_es}{.}

\item{e_df_model}{.}

\item{e_r2}{.}

\item{e_df_effect}{.}

\item{power}{.}

\item{n}{.}

\item{sig.level}{.}

\item{alternative}{.}

\item{plot_contour}{.}

\item{plot_escurve}{.}

\item{plot_ncurve}{.}

\item{plot_log}{.}

\item{plot_palette}{.}

\item{covs}{.}

\item{factors}{.}

\item{factors_list}{.}

\item{covs_order}{.}

\item{factors_order}{.}

\item{mixed_order}{.}

\item{eta}{.}

\item{f}{.}

\item{eta_df_error}{.}

\item{epsilon}{.}

\item{omega}{.}

\item{gpower}{.}

\item{f2}{.}

\item{use}{not used in R}

\item{ncp_type}{What type of non-centrality parameter (NCP) should be used.
The effect size is always transformed into  a Cohen's \code{f2}, that is
multiplied by the N to  which is estimated based on the degrees of freedom
such that \code{N=df+edf+1}. \code{model} defines \code{df} as the model
degrees of freedom. This is the method used by \code{G*Power} software.
\code{liberal} uses the effect \code{df}. \code{strict} uses only the error
df (df=0).}

\item{rx}{a vector of strings naming the columns from \code{data}
containing the correlations among independent variables}

\item{plot_x}{.}

\item{plot_y}{.}

\item{plot_custom_labels}{.}

\item{plot_z}{.}

\item{plot_x_from}{.}

\item{plot_x_to}{.}

\item{plot_z_lines}{.}

\item{plot_z_value}{.}

\item{plot_to_table}{.}

\item{explain}{.}

\item{.caller}{.}

\item{.interface}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$intro} \tab \tab \tab \tab \tab a html \cr
  \code{results$extrainfo} \tab \tab \tab \tab \tab a html \cr
  \code{results$issues} \tab \tab \tab \tab \tab a html \cr
  \code{results$powertab} \tab \tab \tab \tab \tab a table \cr
  \code{results$effectsize} \tab \tab \tab \tab \tab a table \cr
  \code{results$powerbyes} \tab \tab \tab \tab \tab a table \cr
  \code{results$plotnotes} \tab \tab \tab \tab \tab a html \cr
  \code{results$powerContour} \tab \tab \tab \tab \tab an image \cr
  \code{results$powerEscurve} \tab \tab \tab \tab \tab an image \cr
  \code{results$powerNcurve} \tab \tab \tab \tab \tab an image \cr
  \code{results$powerCustom} \tab \tab \tab \tab \tab an image \cr
  \code{results$customnotes} \tab \tab \tab \tab \tab a html \cr
  \code{results$customtable} \tab \tab \tab \tab \tab a table \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$powertab$asDF}

\code{as.data.frame(results$powertab)}
}
\description{
Something here
}
