% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pamlmed.h.R
\name{pamlmed}
\alias{pamlmed}
\title{Mediation}
\usage{
pamlmed(
  aim = "n",
  mode = "medsimple",
  a = 0.3,
  b = 0.3,
  cprime = 0,
  model_type = "twomeds",
  a1 = "",
  b1 = "",
  a2 = "",
  b2 = "",
  a3 = "",
  b3 = "",
  d1 = "",
  d2 = "",
  r12 = "",
  r13 = "",
  r23 = "",
  cprime2 = 0,
  power = 0.9,
  n = 100,
  sig.level = 0.05,
  alternative = "two.sided",
  test = "joint",
  mcR = 1000,
  parallel = FALSE,
  set_seed = FALSE,
  seed = 42,
  sim_ttest = FALSE,
  table_pwbyn = TRUE,
  plot_ncurve = FALSE,
  plot_log = FALSE,
  plot_palette = "viridis",
  plot_x = "none",
  plot_y = "none",
  plot_custom_labels = FALSE,
  plot_z = "none",
  plot_x_from = 0,
  plot_x_to = 0,
  plot_z_lines = 1,
  plot_z_value = list(),
  plot_to_table = FALSE,
  test_c = FALSE,
  explain = FALSE,
  diagram = TRUE,
  .interface = "jamovi",
  .caller = "mediation"
)
}
\arguments{
\item{aim}{The aim of the analysis: \code{n} (default) for sample size,
\code{power} to estimate power, \code{es} for effect size (correlation)}

\item{mode}{.}

\item{a}{The expected standardized effect of the independent variable on
the mediator}

\item{b}{The expected standardized effect of the independent variable on
the mediator}

\item{cprime}{The expected standardized effect of the independent variable
on the mediator}

\item{model_type}{.}

\item{a1}{The expected standardized effect of the independent variable on
mediator 1}

\item{b1}{The expected standardized effect of the independent variable on
mediator 1}

\item{a2}{The expected standardized effect of the independent variable on
mediator 1}

\item{b2}{The expected standardized effect of the independent variable on
mediator 1}

\item{a3}{The expected standardized effect of the independent variable on
mediator 1}

\item{b3}{The expected standardized effect of the independent variable on
mediator 1}

\item{d1}{The expected standardized effect of the independent variable on
mediator 1}

\item{d2}{The expected standardized effect of the independent variable on
mediator 1}

\item{r12}{The expected standardized effect of the independent variable on
mediator 1}

\item{r13}{The expected standardized effect of the independent variable on
mediator 1}

\item{r23}{The expected standardized effect of the independent variable on
mediator 1}

\item{cprime2}{The expected standardized effect of the independent variable
on the mediator}

\item{power}{Minimal desired power}

\item{n}{Sample size}

\item{sig.level}{Type I error rate (significance cut-off or alpha)}

\item{alternative}{.}

\item{test}{.}

\item{mcR}{.}

\item{parallel}{.}

\item{set_seed}{.}

\item{seed}{.}

\item{sim_ttest}{.}

\item{table_pwbyn}{.}

\item{plot_ncurve}{.}

\item{plot_log}{.}

\item{plot_palette}{.}

\item{plot_x}{.}

\item{plot_y}{.}

\item{plot_custom_labels}{.}

\item{plot_z}{.}

\item{plot_x_from}{.}

\item{plot_x_to}{.}

\item{plot_z_lines}{.}

\item{plot_z_value}{.}

\item{plot_to_table}{.}

\item{test_c}{.}

\item{explain}{.}

\item{diagram}{.}

\item{.interface}{Used for internal purposes}

\item{.caller}{Used for internal purposes}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$intro} \tab \tab \tab \tab \tab a html \cr
  \code{results$extrainfo} \tab \tab \tab \tab \tab a html \cr
  \code{results$issues} \tab \tab \tab \tab \tab a html \cr
  \code{results$diagram} \tab \tab \tab \tab \tab an image \cr
  \code{results$powertab} \tab \tab \tab \tab \tab a table \cr
  \code{results$effectsize} \tab \tab \tab \tab \tab a table \cr
  \code{results$powerbyn} \tab \tab \tab \tab \tab a table \cr
  \code{results$powerxy} \tab \tab \tab \tab \tab a table \cr
  \code{results$plotnotes} \tab \tab \tab \tab \tab a html \cr
  \code{results$powerNcurve} \tab \tab \tab \tab \tab an image \cr
  \code{results$powerCustom} \tab \tab \tab \tab \tab an image \cr
  \code{results$customnotes} \tab \tab \tab \tab \tab a html \cr
  \code{results$customtable} \tab \tab \tab \tab \tab a table \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$powertab$asDF}

\code{as.data.frame(results$powertab)}
}
\description{
Something here
}
