% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pamlmixed.h.R
\name{pamlmixed}
\alias{pamlmixed}
\title{Linear Mixed Models}
\usage{
pamlmixed(
  aim = "n",
  find = "k",
  code = "",
  fonts = "small",
  toggle = FALSE,
  run,
  sigma = 1,
  power = 0.9,
  sig.level = 0.05,
  mcR = 500,
  parallel = TRUE,
  set_seed = FALSE,
  seed = 42,
  .interface = "jamovi",
  .caller = "pamlmixed",
  clusterpars = list(),
  var_type = list()
)
}
\arguments{
\item{aim}{The aim of the analysis: \code{n} (default) for sample size,
\code{power} to estimate power}

\item{find}{The aim of the analysis: \code{n} (default) for sample size,
\code{power} to estimate power}

\item{code}{.}

\item{fonts}{.}

\item{toggle}{.}

\item{run}{.}

\item{sigma}{Residual variance}

\item{power}{Minimal desired power}

\item{sig.level}{Type I error rate (significance cut-off or alpha)}

\item{mcR}{Number of repetitions for Monte Carlo method}

\item{parallel}{Logical: should parallel computing be used for the Monte
Carlo method}

\item{set_seed}{not used in R}

\item{seed}{not used in R}

\item{.interface}{Used for internal purposes}

\item{.caller}{Used for internal purposes}

\item{clusterpars}{.}

\item{var_type}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$intro} \tab \tab \tab \tab \tab a html \cr
  \code{results$extrainfo} \tab \tab \tab \tab \tab a html \cr
  \code{results$issues} \tab \tab \tab \tab \tab a html \cr
  \code{results$initnotes} \tab \tab \tab \tab \tab a html \cr
  \code{results$infotab} \tab \tab \tab \tab \tab a table \cr
  \code{results$powertab} \tab \tab \tab \tab \tab a table \cr
  \code{results$plotnotes} \tab \tab \tab \tab \tab a html \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$infotab$asDF}

\code{as.data.frame(results$infotab)}
}
\description{
Something here
}
