% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pamlmixed.f.R
\name{pamlmixed}
\alias{pamlmixed}
\title{Linear Mixed Models}
\usage{
pamlmixed(
  aim = "n",
  find = "n",
  syntax = NULL,
  clusterpars = list(),
  categorical = list(),
  sigma2 = 1,
  power = 0.9,
  sig.level = 0.05,
  algo = "mc",
  mcR = 500,
  parallel = TRUE,
  set_seed = FALSE,
  seed = 42
)
}
\arguments{
\item{aim}{The aim of the analysis: \code{n} (default) sample size,
\code{power} to estimate power}

\item{find}{When \code{aim='n'}, indicates whether to find number of clusters \code{find='k'} or number of cases within each cluster \code{find='n'} (default).}

\item{syntax}{The model to be analysed with possible options}

\item{clusterpars}{A named list of the form `list(cluster1=c(n=n1,k=k1))`, where `cluster1` is the name of the clustering variable
in the model, `n1` is the expcted number of cases within each cluster, and `k1` is the expcted number of clusters. if \code{aim=n}, `n1` is
used as starting point for sample size. If \code{aim=clusters}, `k1` is used as starting point for number of clusters.}

\item{categorical}{A named list of the form `list(varname1=x1,varname2=x2)`, specifying which variable is categorical and the number of levels (x). 
Any variable in the model not mentioned in `categorical ` is assumed to be numeric.}

\item{sigma2}{Residual variance}

\item{power}{Minimal desired power}

\item{sig.level}{Type I error rate (significance cut-off or alpha)}

\item{algo}{The algorithm to use: `mc` (default) for Monte Carlo simulation, `raw` for raw approximation based on Chi-squared (fast but not very accurate)}

\item{mcR}{Number of repetitions for Monte Carlo method}

\item{parallel}{Logical: should parallel computing be used for the Monte
Carlo method}

\item{seed}{the seed for Monte Carlo simulations, default=42.}

\item{Number}{of repetitions for Monte Carlo method}

\item{.interface}{Used for internal purposes}

\item{.caller}{Used for internal purposes}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$intro} \tab \tab \tab \tab \tab a html \cr
  \code{results$extrainfo} \tab \tab \tab \tab \tab a html \cr
  \code{results$issues} \tab \tab \tab \tab \tab a html \cr
  \code{results$initnotes} \tab \tab \tab \tab \tab a html \cr
  \code{results$infotab} \tab \tab \tab \tab \tab a table \cr
  \code{results$powertab} \tab \tab \tab \tab \tab a table \cr
  \code{results$plotnotes} \tab \tab \tab \tab \tab a html \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$infotab$asDF}

\code{as.data.frame(results$infotab)}
}
\description{
power analysis for linear mixed model
}
