% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pamlttest.h.R
\name{pamlttest}
\alias{pamlttest}
\title{T-Test}
\usage{
pamlttest(
  .caller = "ttest",
  aim = "n",
  mode = "ttestind",
  ttestind_es = 0.5,
  ttestind_n = 20,
  ttestind_nratio = 1,
  ttestpaired_es = 0.5,
  ttestpaired_n = 20,
  ttestone_es = 0.5,
  ttestone_n = 20,
  is_equi = FALSE,
  equi_limit = 0.1,
  power = 0.9,
  sig.level = 0.05,
  alternative = "two.sided",
  plot_contour = FALSE,
  plot_escurve = FALSE,
  plot_ncurve = FALSE,
  plot_log = FALSE,
  plot_x = "none",
  plot_y = "none",
  plot_custom_labels = FALSE,
  plot_z = "none",
  plot_x_from = 0,
  plot_x_to = 0,
  plot_z_lines = 0,
  plot_z_value = list(),
  plot_to_table = FALSE
)
}
\arguments{
\item{.caller}{.}

\item{aim}{.}

\item{mode}{.}

\item{ttestind_es}{.}

\item{ttestind_n}{.}

\item{ttestind_nratio}{.}

\item{ttestpaired_es}{.}

\item{ttestpaired_n}{.}

\item{ttestone_es}{.}

\item{ttestone_n}{.}

\item{is_equi}{.}

\item{equi_limit}{.}

\item{power}{.}

\item{sig.level}{.}

\item{alternative}{.}

\item{plot_contour}{.}

\item{plot_escurve}{.}

\item{plot_ncurve}{.}

\item{plot_log}{.}

\item{plot_x}{.}

\item{plot_y}{.}

\item{plot_custom_labels}{.}

\item{plot_z}{.}

\item{plot_x_from}{.}

\item{plot_x_to}{.}

\item{plot_z_lines}{.}

\item{plot_z_value}{.}

\item{plot_to_table}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$intro} \tab \tab \tab \tab \tab a html \cr
  \code{results$issues} \tab \tab \tab \tab \tab a html \cr
  \code{results$powertab} \tab \tab \tab \tab \tab a table \cr
  \code{results$powerbyes} \tab \tab \tab \tab \tab a table \cr
  \code{results$plotnotes} \tab \tab \tab \tab \tab a html \cr
  \code{results$powerContour} \tab \tab \tab \tab \tab an image \cr
  \code{results$powerEscurve} \tab \tab \tab \tab \tab an image \cr
  \code{results$powerNcurve} \tab \tab \tab \tab \tab an image \cr
  \code{results$powerCustom} \tab \tab \tab \tab \tab an image \cr
  \code{results$customnotes} \tab \tab \tab \tab \tab a html \cr
  \code{results$customtable} \tab \tab \tab \tab \tab a table \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$powertab$asDF}

\code{as.data.frame(results$powertab)}
}
\description{
Something here
}
